refactor(repo C7): extract docid utilities to engine/repository/docid.ts
Moves normalizeDocid, isDocid, findDocumentByDocid, findSimilarFiles,
and the internal levenshtein helper. matchFilesByGlob stays in
repository.ts (depends on not-yet-extracted getCollectionByName).

Build: clean. Tests: 970/970 pass.
