+++ engine/repository/index.ts
@@ -8,3 +8,4 @@
 
 export * from "./paths.js";
 export * from "./llm-cache.js";
+export * from "./docid.js";
